// src/api/chat-api/get-chats-api.ts
import {BaseApi} from '../base-api';
import {Http, Options} from '../../services/http-service';
import {host} from '../../constants';
import {ErrorResponse} from '../types';

export type ChatsResponse = Chat[];

export type Chat = {
	id: number;
	title: string;
	avatar: string;
	unread_count: number;
	last_message: LastMessage;
};

export type LastMessage = {
	user: ChatUser;
	time: string;
	content: string;
};

export type ChatUser = {
	first_name: string;
	second_name: string;
	avatar: string;
	email: string;
	login: string;
	phone: string;
};

const getChatsApiInstance = new Http(`${host}/api/v2/chats`);

export class GetChatsApi extends BaseApi {
	public async get(options: Options): Promise<ChatsResponse | ErrorResponse> {
		return getChatsApiInstance.get<ChatsResponse | ErrorResponse>('', options)
			.then(data => data);
	}
}